# Remove redundant layout expansions in GRID config