--- a/bundles/AdminBundle/Resources/public/js/pimcore/object/helpers/classTree.js
+++ b/bundles/AdminBundle/Resources/public/js/pimcore/object/helpers/classTree.js
@@ -161,7 +161,7 @@ pimcore.object.helpers.classTree = Class.create({
                         baseNode.appendChild(this.recursiveAddNode(data[keys[i]].children[j], baseNode, brickDescriptor, this.config));
                     }
                     if (data[keys[i]].nodeType == "object") {
-                        baseNode.expand();
+                        baseNode.expand(true);
                     } else {
                         // baseNode.collapse();
                     }
@@ -224,8 +224,6 @@ pimcore.object.helpers.classTree = Class.create({
         };
 
         newNode = this.appendChild(newNode);
-
-        this.expand();
 
         return newNode;
     },