See https://github.com/pimcore/admin-ui-classic-bundle/pull/293#issue-1913221698

FYI: @brusch  @fashxp @skochTC