@dev-pim 2023-11-12
Hello! Great improvement
However, have you tried the method that I indicated in the comment from the related topic?
https://github.com/pimcore/pimcore/issues/12465#issuecomment-1520945714

Set **bufferedRender** to **true** in https://github.com/pimcore/pimcore/blob/f39ea979c2c5bc460c59703e3b61a0d03697ce65/bundles/AdminBundle/Resources/public/js/pimcore/object/helpers/classTree.js#L94

With your changes I get an increase in speed
24.7 seconds => 2.1 seconds
With changing **bufferedRender** to **true** it's a little better
24.7 seconds => 1.1 seconds
At the same time, the nodes remain expanded

Could you, with your expertise, check the option with **bufferedRender: true**, maybe it will be possible to **return this.expand();** and get a greater increase in performance
[https://github.com/pimcore/pimcore/pull/16007/commits/f39ea979c2c5bc460c59703e3b61a0d03697ce65#diff-ac637775f79f55cd116eff06ef361578e91e4f299b7099754cfb153b 3ac88bc8L228](https://github.com/pimcore/pimcore/commit/f39ea979c2c5bc460c59703e3b61a0d03697ce65#diff-ac637775f79f55cd116eff06ef361578e91e4f299b7099754cfb153b3ac88bc8L228)