@dev-pim 2023-11-12
According to my tests, the line with [baseNode.expand(true)](https://github.com/pimcore/pimcore/commit/f39ea979c2c5bc460c59703e3b61a0d03697ce65#diff-ac637775f79f55cd116eff06ef361578e91e4f299b7099754cfb153b3ac88bc8R164) does not give any speed increase.
Only deleting rows using [this.expand()](https://github.com/pimcore/pimcore/commit/f39ea979c2c5bc460c59703e3b61a0d03697ce65#diff-ac637775f79f55cd116eff06ef361578e91e4f299b7099754cfb153b3ac88bc8L228) has an effect