@dev-pim 2023-11-12
If we use **bufferedRender: true** and **remove this.expand()** we get a little more performance

only **removing this.expand()**
24.7 seconds => 2.1 seconds

only **bufferedRender: true**
24.7 seconds => 1.1 seconds

**bufferedRender: true** + **this.expand()**
24.7 seconds => 0.9 seconds
But in this case, the nodes remain collapse and will have to be expande manually, which, it seems to me, may not be entirely convenient for those who are used to seeing them expanded