@skochTC 2023-11-14
> If we use **bufferedRender: true** and **remove this.expand()** we get a little more performance
> 
> only **removing this.expand()** 24.7 seconds => 2.1 seconds
> 
> only **bufferedRender: true** 24.7 seconds => 1.1 seconds
> 
> **bufferedRender: true** + **this.expand()** 24.7 seconds => 0.9 seconds But in this case, the nodes remain collapse and will have to be expande manually, which, it seems to me, may not be entirely convenient for those who are used to seeing them expanded

Hi @dev-pim, thanks for you note. If you see any improvements fell free to extend my changes.

My changes was already merged for Pimcore >=X, see https://github.com/pimcore/admin-ui-classic-bundle/pull/293#issue-1913221698

We did a lot of mersuremtns and for us we was able to reduce the client "overheat" down to this:
> In my test case, I was able to reduce the wait time from an average of 30,619s to 1,434s.
Just by ading a true to the `baseNode.expand();` call

BR, Sebastian