@herbertroth 2023-12-06
Fixed in LTS version.